feat: add driver assignments, add route adding & add base price in trips

+++ src/main/java/mk/route/routemk/web/company/CompanyPerformanceController.java
@@ -43,7 +43,7 @@ public class CompanyPerformanceController {
             if (topSellingRoutesView != null && !topSellingRoutesView.isEmpty()) {
 
             if (topSellingRoutesView.size() > 10) {
-                Long maxPopularity = topSellingRoutesView.stream()
+                Integer maxPopularity = topSellingRoutesView.stream()
                         .max(Comparator.comparing(TopSellingRouteView::getTotalTicketsSold))
                         .get().getTotalTicketsSold();
